Rename table_common.js to people_five_year_percentage_formula.js. Don't want more exports slipping in later, should only be downscaling on the table code
